Update Treasury: # of Tranches & Issuance Interval
Article 21: Change number of tranches from 12 to 10 to match whitepaper v1.2. 

Article 22: Add a minimum interval of 8-12 months. This is a guardrail for a more steady funding schedule, while allowing for some flexibility. 
Though it is unlikely that we would have multiple back-to-back treasury funding proposals, this change would assuage the concerns of the community regarding "inflation shock".