Update: Set Issuance Interval | Article 22
Removed "There must be a minimum interval of 8 months between each successful proposal" and kept the calendar year limit. This was for simplicity. 

Used a calendar year limit instead of a 12-month interval. With a 12-month interval funding gets pushed back a bit every year unless its requested as soon as possible.